split tapelike/familiar/cursor into their own files
the time for the great refactoring is upon me
