+++ src/game/chess/familiar.rs
@@ -2,6 +2,7 @@
 use crate::{coup::rep::Move, interface::play::Play, notation::ben::BEN, types::{log::cursor::Cursor, movesheet::MoveSheet}};
 use super::{action::Action, delim::Delim};
 
+// TODO: port this to the new system.
 #[derive(Debug, Clone)]
 pub struct Familiar<'a> {
     // TODO: Temporarily fixing the types
