+++ src/types/tape/tape_direction.rs
@@ -1,0 +1,16 @@
+
+#[derive(PartialEq, Debug, Clone, Copy, Eq)]
+pub enum TapeDirection {
+    Advancing,
+    Rewinding
+}
+
+impl TapeDirection {
+    pub fn advancing(self) -> bool {
+        self == TapeDirection::Advancing
+    }
+
+    pub fn rewinding(self) -> bool {
+        self == TapeDirection::Rewinding
+    }
+}
