+++ src/types/tape/tapelike.rs
@@ -1,0 +1,19 @@
+use std::range::Range;
+
+use itertools::Itertools;
+
+use crate::{Alter, Alteration};
+
+use super::tape::Tape;
+
+// Covers all the IO operations on the tape, without an explicit read/write head being maintained.
+pub trait Tapelike {
+    type Item;
+
+    fn length(&self) -> usize;
+
+    fn read_address(&self, address: usize) -> Option<&Self::Item>;
+    fn read_range(&self, range: impl Into<Range<usize>>) -> &[Option<Self::Item>];
+    fn write_address(&mut self, address: usize, data: &Self::Item);
+    fn write_range(&mut self, start: usize, data: &[Self::Item]);
+}
